Added more functionality
This commit added text fields and buttons for submitting data in the text fields as well as a button for running the data. The program is still not entirely functioning. I have not been able to find a way to save data from the text boxes for use in other functions, as it gives me an error saying variables must be final.